Fix browser command-line arguments type error
Fixes an error where PLOT-FROM-FILE would attempt to splice the return
value of ENCODE-APPLICATION-OPTIONS as if it were a list when in fact
it is a string.

Since ENCODE-APPLICATION-OPTIONS returns a string that can contain
multiple command-line arguments it is now passed as a command-line
string to UIOP:LAUNCH-PROGRAM so that it can be properly tokenized by
the shell.